+++ R/messages.R
@@ -85,3 +85,7 @@ messages$valueNotPositive <- function(value, propertyName) {
 messages$molWeightErrorMessage <- function(quantityPath) {
   paste0("Unable to retrieve the molecular weight for: ", quantityPath)
 }
+
+messages$illegalCharactersInName <- function(name) {
+  paste0("The name '", name, "' contains illegal characters. Illegal characters are: '", paste0(getOSPSuiteSetting("illegalCharacters"), collapse = ", "), "'.")
+}